third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/dont-report-injected-script.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/console-test.js"></script>

<script>
function newWorld()
{
    testRunner.evaluateScriptInIsolatedWorld(239, "console.log(42);\n//# sourceURL=239.js");
}

function test() {
  TestRunner.debuggerModel.addEventListener(SDK.DebuggerModel.Events.ParsedScriptSource, step2);
  TestRunner.evaluateInPage('newWorld()\n//# sourceURL=foo.js');

  var expectedScriptParsed = 2;
  function step2(event) {
    var script = event.data;
    TestRunner.addResult(script.sourceURL);
    --expectedScriptParsed;
    if (!expectedScriptParsed)
      TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that injected script isn't reported to frontend.
</p>
</body>
</html>
